requirements.txt
asgiref==3.2.10
Django==3.0.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
django-filter==2.4.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
pytest-django==3.9.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
six==1.16.0
